refactoring to speed up local testing loop and, hopefully, github actions loop:
Changes made
New file: MukurtuKernelTestBase.php
The shared abstract base class. It owns:

Common $modules list (the intersection of all three existing bases)
Blazy file.repository synthetic stub + mock (defensive, harmless for non-blazy tests)
All schema/entity/config installs for: sequences, file, protocol, local_contexts, taxonomy, OG, workflow
OG group registration, authenticated role, protocol steward OG role
$this->currentUser, $this->community, $this->protocol fixture creation
getProtocolStewardPermissions() hook for subclasses to add extra OG permissions
Refactored: LocalContextsTestBase.php
~110 lines → 55 lines. No $modules redeclaration needed (its list was identical to the parent's). setUp now only sets $this->manager.

Refactored: MukurtuMediaTestBase.php
~220 lines → 132 lines. Redeclares $modules with its extra modules. setUp handles only the media-specific FieldStorageConfig/MediaType/FieldConfig/cache/vocabulary setup.

Refactored: DictionaryTestBase.php
~290 lines → 185 lines. Redeclares $modules with its large extra set. Overrides getProtocolStewardPermissions() to add dictionary CRUD OG perms. setUp adds comment/node/paragraph/search_api schemas plus NodeType, ParagraphsType, vocabularies, language term.

Updated: phpunit.xml
Added mukurtu_core/tests/src/Kernel to the kernel testsuite (ready for future concrete tests).

Key design note: Since PHP static property overriding means a child's $modules fully replaces the parent's, any subclass needing extra modules must redeclare $modules with the complete list — there's no automatic merging. The deduplication benefit is entirely in setUp() code.
